pyproject.toml
[tool.poetry]
name = "llm-website-generator"
version = "0.1.0"
description = ""
authors = ["Chongrui Zhang"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.55.0"
flask = "^3.1.0"
requests = "^2.32.3"
pillow = "^11.0.0"
selenium = "^4.27.1"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
